+++ build-tools/split.sh
@@ -40,6 +40,7 @@ remote properties
 remote shipment
 remote support
 remote taxes
+remote translation
 remote video
 
 split 'src/Adjustments' adjustments
@@ -58,4 +59,5 @@ split 'src/Properties' properties
 split 'src/Shipment' shipment
 split 'src/Support' support 8683e47dd2dbd15ac2ceac4dcfae405c4b271aff
 split 'src/Taxes' taxes
+split 'src/Translation' translation
 split 'src/Video' video