+++ src/Translation/Tests/TestCase.php
@@ -1,0 +1,52 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Vanilo\Translation\Tests;
+
+use Illuminate\Foundation\Application;
+use Konekt\Concord\ConcordServiceProvider;
+use Orchestra\Testbench\TestCase as Orchestra;
+use Vanilo\Translation\Providers\ModuleServiceProvider as TranslationModule;
+
+abstract class TestCase extends Orchestra
+{
+    public function setUp(): void
+    {
+        parent::setUp();
+
+        $this->setUpDatabase($this->app);
+    }
+
+    protected function getPackageProviders($app): array
+    {
+        return [
+            ConcordServiceProvider::class
+        ];
+    }
+
+    protected function getEnvironmentSetUp($app): void
+    {
+        $app['config']->set('database.default', 'sqlite');
+        $app['config']->set('database.connections.sqlite', [
+            'driver' => 'sqlite',
+            'database' => ':memory:',
+            'prefix' => '',
+        ]);
+    }
+
+    protected function setUpDatabase(Application $app): void
+    {
+        $this->loadMigrationsFrom(__DIR__ . '/migrations');
+        \Artisan::call('migrate', ['--force' => true]);
+    }
+
+    protected function resolveApplicationConfiguration($app): void
+    {
+        parent::resolveApplicationConfiguration($app);
+
+        $app['config']->set('concord.modules', [
+            TranslationModule::class,
+        ]);
+    }
+}